TIP-1098 Save invalid data file to flystystem
